Change data type to double.

+++ Domain/Models/ProjectSetting.cs
@@ -7,10 +7,10 @@ namespace SchedulingTool.Api.Domain.Models
     {
         public long VersionId { get; set; }
         public bool SeparateGroupTask { get; set; }
-        public decimal AssemblyDurationRatio { get; set; }
-        public decimal RemovalDurationRatio { get; set; }
+        public double AssemblyDurationRatio { get; set; }
+        public double RemovalDurationRatio { get; set; }
         public int ColumnWidth { get; set; }
-        public decimal AmplifiedFactor { get; set; }
+        public double AmplifiedFactor { get; set; }
         public bool IncludeYear { get; set; }
         public int StartYear { get; set; }
         public int StartMonth { get; set; }
